VC++ fixes

+++ src/geom_correction_queue.cpp
@@ -41,7 +41,7 @@ namespace phtr
     GeomCorrectionQueue::GeomCorrectionQueue(const GeomCorrectionQueue& orig)
             : n_models_(0)
     {
-        n_models_ = orig.correction_model_.size();
+        n_models_ = static_cast<unsigned short>(orig.correction_model_.size());
         correction_model_.resize(n_models_);
 
         for (size_t i = 0; i < n_models_; ++i)
@@ -59,7 +59,7 @@ namespace phtr
     {
         this->clear();
 
-        n_models_ = orig.correction_model_.size();
+        n_models_ = static_cast<unsigned short>(orig.correction_model_.size());
         correction_model_.resize(n_models_);
 
         for (size_t i = 0; i < n_models_; ++i)
@@ -72,7 +72,7 @@ namespace phtr
 
     void GeomCorrectionQueue::clear()
     {
-        n_models_ = correction_model_.size();
+        n_models_ = static_cast<unsigned short>(correction_model_.size());
 
         for (size_t i = 0; i < n_models_; ++i)
         {
